% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_smooths.R
\name{evaluate_smooths}
\alias{evaluate_smooths}
\title{evaluate smooths at new data}
\usage{
evaluate_smooths(x, newdata)
}
\arguments{
\item{x}{a greta array created with greta.gam::smooths}

\item{newdata}{a dataframe with the same column names and datatypes as that
used to create x, with data at which to evauate the smooths}
}
\value{

}
\description{
Evaluate a set of smooths at new data locations
}
\author{
Nick Golding
}
