% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greta-gam-package.R
\docType{package}
\name{greta.gam}
\alias{greta.gam-package}
\alias{greta.gam}
\title{greta.gam: generalised additive models in greta using mgcv}
\description{
greta.gam is a module for greta that lets you use mgcv's
smoother functions and formula syntax to define smooth terms for use in a
greta model. You can then define your own likelihood to complete the
model, and fit it by MCMC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/greta-dev/greta.gam}
  \item \url{https://greta-dev.github.io/greta.gam/}
  \item Report bugs at \url{https://github.com/greta-dev/greta.gam/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Authors:
\itemize{
  \item Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})
  \item Nick Golding \email{nick.golding.research@gmail.com} (\href{https://orcid.org/0000-0001-8916-5570}{ORCID}) [copyright holder]
  \item David L Miller \email{dave@ninepointeightone.net}
}

}
\keyword{internal}
