% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagam2greta.R
\name{jagam2greta}
\alias{jagam2greta}
\title{Turn a jagam model definition into a greta model definition}
\usage{
jagam2greta(formula, data, newdata, sp = NULL, knots = NULL)
}
\arguments{
\item{formula}{a GAM formula representing the smooth terms, as in
\code{\link[mgcv:gam]{mgcv::gam}}. Only the right hand side fo the formula
will be used.}

\item{data}{a data frame or list containing the covariates required by the
formula. These covariates cannot be greta arrays.}

\item{newdata}{new data to use for predictions etc (this probably doesn't likve here but is supplied for now for testing)}

\item{sp}{an optional vector of smoothing parameters, two per smooth term in
the model, in the same order as the formula. If \code{sp = NULL}, all
smoothing parameters will be learned, otherwise all smoothing parameters
must be specified by the user. The smoothing parameters may either be a
numeric vector or a greta array (which could be a variable).}

\item{knots}{an optional list containing user specified knot values to be
used for basis construction, as in \code{\link[mgcv:gam]{mgcv::gam}} These
knots cannot be greta arrays.}
}
\value{
a \code{list} with the following elements: \code{betas} a greta array for the coefficients to be estimated (with appropriate priors applied), \code{X} design matrix for this model, \code{X_pred} prediction matrix.
}
\description{
Takes a GAM defined by \code{formula} and returns the corresponding \code{greta} model via the power of \code{jagam}. Response variable is generated from dummy data and not used.
}
