% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.gam}
\alias{greta.gam}
\alias{greta.gam-package}
\title{greta.gam: generalised additive models in greta using mgcv}
\description{
greta.gam is a module for greta that lets you use mgcv's smoother
  functions and formula syntax to define smooth terms for use in a greta
  model. You can then define your own likelihood to complete the model, and
  fit it by MCMC.
}
